move version_bump file to apps_ci
