+++ catalog_reader/scripts/apps_hashes.py
@@ -8,8 +8,8 @@ import yaml
 from apps_exceptions import CatalogDoesNotExist, ValidationErrors
 from catalog_reader.dev_directory import get_ci_development_directory
 from catalog_reader.library import get_hashes_of_base_lib_versions
+from apps_ci.version_bump import is_valid_bump_type, bump_version, rename_versioned_dir
 from catalog_reader.names import get_library_path, get_library_hashes_path, get_base_library_dir_name_from_version
-from catalog_reader.version_bump import is_valid_bump_type, bump_version, rename_versioned_dir
 
 
 def update_catalog_hashes(catalog_path: str, bump_type: str | None = None) -> None:
